Move AcquisitionData to its own file.
